Complete 4th lesson
Added a field User for the class Message.
Implemented the method getAuthority (GrantedAuthority) for the interface Role.
A class User was implemented by the interface UserDetails.
Added a class UserService that is implementing UserDetailsService.
Changed the class WebSecurityConfig, exactly, a method  configure (alter from jdbc auth to custom auth).
Updated a method add for the class MainController in view updating domains.
 Updated a view for main.mustache.