Complete 9th lesson
Updated a pom file: add a mail dependency.
Update the file properties: add the mail properties.
Updated a class User: add two field email and activationCode.
Added a service EmailSenderService that handler the tasks related with the sending letters.
Updated a controllers and it's the dependency classes.
Updated the views.